+++ R/explain.R
@@ -15,6 +15,7 @@
 #' @param label character - the name of the model. By default it's extracted from the 'class' attribute of the model
 #' @param verbose if TRUE (default) then diagnostic messages will be printed
 #' @param precalculate if TRUE (default) then \code{predicted_values} and \code{residual} are calculated when explainer is created.
+#' @param colorize if TRUE (default) then \code{WARNINGS}, \code{ERRORS} and \code{NOTES} are colorized. Will work only in the R console.
 #' This will happen also if \code{verbose} is TRUE. Set both \code{verbose} and \code{precalculate} to FALSE to omit calculations.
 #'
 #' @return An object of the class \code{explainer}.
@@ -58,7 +59,7 @@
 #' aps_lm <- explain(aps_lm_model4, data = apartments, label = "model_4v", y = apartments$m2.price,
 #'                                    predict_function = predict)
 #'
-#'  \dontrun{
+#' \dontrun{
 #' # more complex model
 #' library("randomForest")
 #' aps_rf_model4 <- randomForest(m2.price ~., data = apartments)
@@ -69,15 +70,27 @@
 
 #' @export
 #' @rdname explain
-explain.default <- function(model, data = NULL, y = NULL, predict_function = NULL, residual_function = NULL, ..., label = NULL, verbose = TRUE, precalculate = TRUE) {
+explain.default <- function(model, data = NULL, y = NULL, predict_function = NULL, residual_function = NULL, ...,
+                            label = NULL, verbose = TRUE, precalculate = TRUE, colorize = TRUE) {
   verbose_cat("Preparation of a new explainer is initiated\n", verbose = verbose)
+
+  # set colors
+  if (colorize) {
+    color_codes <- list(yellow_start = "\033[33m", yellow_end = "\033[39m",
+                        red_start = "\033[31m", red_end = "\033[39m",
+                        green_start = "\033[32m", green_end = "\033[39m")
+  } else {
+    color_codes <- list(yellow_start = "", yellow_end = "",
+                        red_start = "", red_end = "",
+                        green_start = "", green_end = "")
+  }
 
   # REPORT: checks for model label
   if (is.null(label)) {
     # label not specified
     # try to extract something
     label <- tail(class(model), 1)
-    verbose_cat("  -> model label       : ", label, " (\033[33mdefault\033[39m)\n", verbose = verbose)
+    verbose_cat("  -> model label       : ", label, " (",color_codes$yellow_start,"default",color_codes$yellow_end,")\n", verbose = verbose)
   } else {
     verbose_cat("  -> model label       : ", label, "\n", verbose = verbose)
   }
@@ -91,7 +104,7 @@ explain.default <- function(model, data = NULL, y = NULL, predict_function = NUL
       data <- possible_data
       verbose_cat("  -> data              : ", nrow(data), " rows ", ncol(data), " cols (\033[33mextracted from the model\033[39m)\n", verbose = verbose)
     } else {
-      verbose_cat("  -> no data avaliable! (\033[31mWARNING\033[39m)\n", verbose = verbose)
+      verbose_cat("  -> no data avaliable! (",color_codes$red_start,"WARNING",color_codes$red_end,")\n", verbose = verbose)
     }
   } else {
     verbose_cat("  -> data              : ", nrow(data), " rows ", ncol(data), " cols \n", verbose = verbose)
@@ -105,24 +118,24 @@ explain.default <- function(model, data = NULL, y = NULL, predict_function = NUL
   # REPORT: checks for y
   if (is.null(y)) {
     # y not specified
-    verbose_cat("  -> target variable   :  not specified! (\033[31mWARNING\033[39m)\n", verbose = verbose)
+    verbose_cat("  -> target variable   :  not specified! (",color_codes$red_start,"WARNING",color_codes$red_end,")\n", verbose = verbose)
   } else {
     if (is.data.frame(y)) {
       y <- unlist(y, use.names = FALSE)
-      verbose_cat("  -> target variable   :  Argument 'y' was a data frame. Converted to a vector. (\033[31mWARNING\033[39m)\n", verbose = verbose)
+      verbose_cat("  -> target variable   :  Argument 'y' was a data frame. Converted to a vector. (",color_codes$red_start,"WARNING",color_codes$red_end,")\n", verbose = verbose)
     }
     verbose_cat("  -> target variable   : ", length(y), " values \n", verbose = verbose)
     if (length(y) != nrow(data)) {
-      verbose_cat("  -> target variable   :  length of 'y; is different than number of rows in 'data' (\033[31mWARNING\033[39m) \n", verbose = verbose)
+      verbose_cat("  -> target variable   :  length of 'y; is different than number of rows in 'data' (",color_codes$red_start,"WARNING",color_codes$red_end,") \n", verbose = verbose)
     }
     if ((is.factor(y) | is.character(y))) {
-      verbose_cat("  -> target variable   :  Please note that 'y' is a factor.  (\033[31mWARNING\033[39m)\n", verbose = verbose)
+      verbose_cat("  -> target variable   :  Please note that 'y' is a factor.  (",color_codes$red_start,"WARNING",color_codes$red_end,")\n", verbose = verbose)
       verbose_cat("  -> target variable   :  Consider changing the 'y' to a logical or numerical vector.\n", verbose = verbose)
       verbose_cat("  -> target variable   :  Otherwise I will not be able to calculate residuals or loss function.\n", verbose = verbose)
     }
 
     if (!is.null(data) & is_y_in_data(data, y)) {
-      verbose_cat("  -> data              :  A column identical to the target variable `y` has been found in the `data`.  (\033[31mWARNING\033[39m)\n", verbose = verbose)
+      verbose_cat("  -> data              :  A column identical to the target variable `y` has been found in the `data`.  (",color_codes$red_start,"WARNING",color_codes$red_end,")\n", verbose = verbose)
       verbose_cat("  -> data              :  It is highly recommended to pass `data` without the target variable column\n", verbose = verbose)
     }
   }
@@ -137,9 +150,9 @@ explain.default <- function(model, data = NULL, y = NULL, predict_function = NUL
     class(yhat_functions) = "character"
     matching_yhat <- intersect(paste0("yhat.", class(model)), yhat_functions)
     if (length(matching_yhat) == 0) {
-      verbose_cat("  -> predict function  : yhat.default will be used (\033[33mdefault\033[39m)\n", verbose = verbose)
+      verbose_cat("  -> predict function  : yhat.default will be used (",color_codes$yellow_start,"default",color_codes$yellow_end,")\n", verbose = verbose)
     } else {
-      verbose_cat("  -> predict function  : ",matching_yhat[1]," will be used (\033[33mdefault\033[39m)\n", verbose = verbose)
+      verbose_cat("  -> predict function  : ",matching_yhat[1]," will be used (",color_codes$yellow_start,"default",color_codes$yellow_end,")\n", verbose = verbose)
     }
   } else {
     verbose_cat("  -> predict function  : ", deparse(substitute(predict_function)), "\n", verbose = verbose)
@@ -150,10 +163,10 @@ explain.default <- function(model, data = NULL, y = NULL, predict_function = NUL
     y_hat <- try(predict_function(model, data), silent = TRUE)
     if (class(y_hat) == "try-error") {
       y_hat <- NULL
-      verbose_cat("  -> predicted values  :  the predict_function returns an error when executed (\033[31mWARNING\033[39m) \n", verbose = verbose)
+      verbose_cat("  -> predicted values  :  the predict_function returns an error when executed (",color_codes$red_start,"WARNING",color_codes$red_end,") \n", verbose = verbose)
     } else {
       if ((is.factor(y_hat) | is.character(y_hat))) {
-        verbose_cat("  -> predicted values  :  factor (\033[31mWARNING\033[39m) with levels: ", paste(unique(y_hat), collapse = ", "), "\n", verbose = verbose)
+        verbose_cat("  -> predicted values  :  factor (",color_codes$red_start,"WARNING",color_codes$red_end,") with levels: ", paste(unique(y_hat), collapse = ", "), "\n", verbose = verbose)
       } else {
         verbose_cat("  -> predicted values  :  numerical, min = ", min(y_hat), ", mean = ", mean(y_hat), ", max = ", max(y_hat), " \n", verbose = verbose)
       }
@@ -168,7 +181,7 @@ explain.default <- function(model, data = NULL, y = NULL, predict_function = NUL
       residual_function <- function(model, data, y) {
         y - predict_function(model, data)
       }
-      verbose_cat("  -> residual function :  difference between y and yhat (\033[33mdefault\033[39m)\n", verbose = verbose)
+      verbose_cat("  -> residual function :  difference between y and yhat (",color_codes$yellow_start,"default",color_codes$yellow_end,")\n", verbose = verbose)
     }
   } else {
     verbose_cat("  -> residual function : ", deparse(substitute(residual_function)), "\n", verbose = verbose)
@@ -179,10 +192,10 @@ explain.default <- function(model, data = NULL, y = NULL, predict_function = NUL
     residuals <- try(residual_function(model, data, y), silent = TRUE)
     if (class(residuals) == "try-error") {
       residuals <- NULL
-      verbose_cat("  -> residuals         :  the residual_function returns an error when executed (\033[31mWARNING\033[39m) \n", verbose = verbose)
+      verbose_cat("  -> residuals         :  the residual_function returns an error when executed (",color_codes$red_start,"WARNING",color_codes$red_end,") \n", verbose = verbose)
     } else {
       if ((is.factor(residuals) | is.character(residuals))) {
-        verbose_cat("  -> residuals         :  factor (\033[31mWARNING\033[39m) with levels: ", paste(unique(residuals), collapse = ", "), "\n", verbose = verbose)
+        verbose_cat("  -> residuals         :  factor (",color_codes$red_start,"WARNING",color_codes$red_end,") with levels: ", paste(unique(residuals), collapse = ", "), "\n", verbose = verbose)
       } else {
         verbose_cat("  -> residuals         :  numerical, min = ", min(residuals), ", mean = ", mean(residuals), ", max = ", max(residuals), " \n", verbose = verbose)
       }
@@ -203,7 +216,7 @@ explain.default <- function(model, data = NULL, y = NULL, predict_function = NUL
   class(explainer) <- "explainer"
 
   # everything went OK
-  verbose_cat("\033[32mA new explainer has been created!\033[39m\n", verbose = verbose)
+  verbose_cat("",color_codes$green_start,"A new explainer has been created!",color_codes$green_end,"\n", verbose = verbose)
   explainer
 }